msvc-integrator-service/requirements.txt
flask==2.3.3
pika==1.3.2
pyjwt==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-flask-exporter==0.22.4
healthcheck